[mapleall]Internal commit msg:     [ICSL] Clean Code 4

+++ src/mapleall/maple_ir/src/mir_nodes.cpp
@@ -291,14 +291,14 @@ BlockNode *BlockNode::CloneTreeWithFreqs(MapleAllocator &allocator,
     FreqType oldFreq = fromFreqs[GetStmtID()];
     FreqType newFreq;
     if (updateOp & kUpdateUnrollRemainderFreq) {
-      newFreq = denom > 0 ? (oldFreq * numer % denom) : oldFreq;
+      newFreq = denom > 0 ? (oldFreq * numer % static_cast<int64_t>(denom)) : oldFreq;
     } else {
-      newFreq = numer == 0 ? 0 : (denom > 0 ? (oldFreq * numer / denom) : oldFreq);
+      newFreq = numer == 0 ? 0 : (denom > 0 ? (oldFreq * numer / static_cast<int64_t>(denom)) : oldFreq);
     }
     toFreqs[nnode->GetStmtID()] = (newFreq > 0 || (numer == 0)) ? newFreq : 1;
     if ((updateOp & kUpdateOrigFreq) != 0) {  // upateOp & 1 : update from
       int64_t left = static_cast<int64_t>(((oldFreq - newFreq) > 0 || (oldFreq == 0)) ? (oldFreq - newFreq) : 1);
-      fromFreqs[GetStmtID()] = static_cast<uint64_t>(left);
+      fromFreqs[GetStmtID()] = static_cast<int64_t>(left);
     }
   }
   for (auto &stmt : stmtNodeList) {
@@ -322,15 +322,15 @@ BlockNode *BlockNode::CloneTreeWithFreqs(MapleAllocator &allocator,
         FreqType oldFreq = fromFreqs[stmt.GetStmtID()];
         FreqType newFreq;
         if ((updateOp & kUpdateUnrollRemainderFreq) != 0) {
-          newFreq = denom > 0 ? (oldFreq * numer % denom) : oldFreq;
+          newFreq = denom > 0 ? (oldFreq * numer % static_cast<int64_t>(denom)) : oldFreq;
         } else {
-          newFreq = numer == 0 ? 0 : (denom > 0 ? (oldFreq * numer / denom) : oldFreq);
+          newFreq = numer == 0 ? 0 : (denom > 0 ? (oldFreq * numer / static_cast<int64_t>(denom)) : oldFreq);
         }
         toFreqs[newStmt->GetStmtID()] =
-            (newFreq > 0 || oldFreq == 0 || numer == 0) ? static_cast<uint64_t>(newFreq) : 1;
+            (newFreq > 0 || oldFreq == 0 || numer == 0) ? static_cast<int64_t>(newFreq) : 1;
         if ((updateOp & kUpdateOrigFreq) != 0) {
           FreqType left = static_cast<int64_t>(((oldFreq - newFreq) > 0 || oldFreq == 0) ? (oldFreq - newFreq) : 1);
-          fromFreqs[stmt.GetStmtID()] = static_cast<uint64_t>(left);
+          fromFreqs[stmt.GetStmtID()] = static_cast<int64_t>(left);
         }
       }
     }
